chore: BridgeContainer refactor
