+++ apps/tangle-dapp/src/data/bridge/useHyperlaneQuote.ts
@@ -20,10 +20,26 @@ export type HyperlaneQuoteProps = {
   recipientAddress: string;
 };
 
-export const getHyperlaneQuote = async (props: HyperlaneQuoteProps | null) => {
+export type HyperlaneQuote = {
+  fees: {
+    local: {
+      amount: bigint;
+      symbol: string;
+    };
+    interchain: {
+      amount: bigint;
+      symbol: string;
+    };
+  };
+};
+
+export const getHyperlaneQuote = async (
+  props: HyperlaneQuoteProps | null,
+): Promise<HyperlaneQuote | null> => {
   try {
     if (!props) {
-      throw new Error('Hyperlane quote props are required');
+      console.error('Hyperlane quote props are required');
+      return null;
     }
 
     const {
@@ -38,7 +54,8 @@ export const getHyperlaneQuote = async (props: HyperlaneQuoteProps | null) => {
     const warpCore = getHyperlaneWarpCore();
 
     if (!warpCore) {
-      throw new Error('Hyperlane warp core not found');
+      console.error('Hyperlane warp core not found');
+      return null;
     }
 
     const tokenToBridge =
@@ -50,8 +67,10 @@ export const getHyperlaneQuote = async (props: HyperlaneQuoteProps | null) => {
       getHyperlaneChainName(sourceTypedChainId),
       tokenToBridge,
     );
+
     if (!hyperlaneToken) {
-      throw new Error('Hyperlane token not found');
+      console.error('Hyperlane token not found');
+      return null;
     }
 
     const originTokenAmount = hyperlaneToken.amount(amount);
@@ -62,6 +81,7 @@ export const getHyperlaneQuote = async (props: HyperlaneQuoteProps | null) => {
         originTokenAmount,
         destination,
       });
+
     if (!isCollateralSufficient) {
       const balance = await fetchEvmTokenBalance(
         senderAddress,
@@ -77,6 +97,7 @@ export const getHyperlaneQuote = async (props: HyperlaneQuoteProps | null) => {
       });
 
       console.error('Insufficient destination collateral');
+      return null;
     }
 
     const errors = await warpCore.validateTransfer({
@@ -87,7 +108,8 @@ export const getHyperlaneQuote = async (props: HyperlaneQuoteProps | null) => {
     });
 
     if (errors) {
-      throw new Error('Invalid transfer parameters');
+      console.error('Invalid transfer parameters');
+      return null;
     }
 
     const fees = await warpCore.estimateTransferRemoteFees({
@@ -111,10 +133,10 @@ export const getHyperlaneQuote = async (props: HyperlaneQuoteProps | null) => {
   } catch (e: unknown) {
     if (axios.isAxiosError(e) && e.response) {
       console.error('Hyperlane quote error:', e.response.data);
-      throw e.response.data;
+      return null;
     }
     console.error('Error fetching hyperlane quote:', e);
-    throw e;
+    return null;
   }
 };
 
@@ -124,5 +146,6 @@ export const useHyperlaneQuote = (props: HyperlaneQuoteProps | null) => {
     queryFn: () => getHyperlaneQuote(props),
     enabled: false,
     refetchOnWindowFocus: false,
+    initialData: null,
   });
 };
